server/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
SQLAlchemy==1.4.23
python-dotenv==0.19.0
bcrypt==3.2.0
PyJWT==2.1.0 